requirements.txt
pandas>=2.0.0
openpyxl>=3.0.0
pywebview>=4.0.0
pyarrow>=10.0.0
